removed CompletedFuture
